NBTree index integrated, tested working with enclave.
The nbtree index used in SGX is integrated on the prototype and currently
supports indentation of a single string column. The prototype seems to be working
correctly and without memory leaks. It was tested with large scripts, with 10000 rows and different comparison operators. Furthermore, it was tested with
valgrind.

The only limitation in terms of functionality is the lack of support of a backward scan which implies that the index can not search for < or <= operators. This is only a prototype limitation and could be completely integrated. As such, the nbtree index correctly supports =, >. and >=.